pci: fix layerscape
On the LX2160A rev 1, SVR_SOC_VER() returns 0x873610, and this causes
the wrong PCIe fdt fixup code to be called, resulting in all PCIe
devices remaining disabled in the DT passed to the kernel.  Fix this
by detecting the ID and using the correct gen4 layerscape PCIe code.

Signed-off-by: Russell King <rmk@armlinux.org.uk>
Signed-off-by: Grant Likely <grant.likely@arm.com>